+++ avro_to_python/classes/field.py
@@ -14,12 +14,14 @@ class Field(object):
     array_item_type = None
     union_types = None
     map_type = None
+    is_enum = False
 
     def __init__(self, name: str, fieldtype: str, avrotype: str=None,
                  default: Union[int, str, float, dict, None]=None,
                  reference_name: str=None, reference_namespace: str=None,
                  array_item_type: 'Field'=None, union_types: List['Field']=[],
-                 map_type: 'Field'=None
+                 map_type: 'Field'=None,
+                 is_enum: bool=None,
                  ):
         self.name = name
         self.fieldtype = fieldtype
@@ -30,6 +32,7 @@ class Field(object):
         self.array_item_type = array_item_type
         self.union_types = union_types
         self.map_type = map_type
+        self.is_enum = is_enum
 
     def __eq__(self, other: Union['Field', str]):
         if isinstance(other, Field):