+++ avro_to_python/utils/avro/primitive_types.py
@@ -21,3 +21,24 @@ PRIMITIVE_TYPE_MAP = {
     'bytes': 'bytes',
     'string': 'str'
 }
+
+# Equivalences for isinstance checks as numbers in JSON can omit fractional part (parsed as int)
+IS_INSTANCE_PRIMITIVE_TYPE_MAP = {
+    'null': 'None',
+    'boolean': 'bool',
+    'int': 'int',
+    'long': 'int',
+    'float': '(float, int)',
+    'double': '(float, int)',
+    'bytes': 'bytes',
+    'string': 'str'
+}
+
+IS_INSTANCE_PRIMITIVE_TYPE_EQ_MAP = {
+    'None': 'None',
+    'bool': 'bool',
+    'int': 'int',
+    'float': '(float, int)',
+    'bytes': 'bytes',
+    'str': 'str'
+}