+++ tests/reader/avro/test_reader.py
@@ -150,6 +150,7 @@ class AvroReaderTests(unittest.TestCase):
                 'symbols': ['VANILLA', 'CHOCOLATE', 'STRAWBERRY']
             },
             'imports': [],
+            'aliased_imports': [],
             'fields': {},
             'symbols': ['VANILLA', 'CHOCOLATE', 'STRAWBERRY'],
             'default': None
@@ -358,7 +359,7 @@ class AvroReaderTests(unittest.TestCase):
         # should have 1 field
         self.assertEqual(
             len(obj.children['records'].files['RecordWithMap'].fields),
-            4
+            6
         )
 
         # field should be of type map