+++ tests/writer/test_compiled_files.py
@@ -1,7 +1,9 @@
 """ tests for the reader and reader utils """
-
+import glob
+import importlib
 import json
 import os
+import pathlib
 import shutil
 import sys
 import unittest
@@ -29,9 +31,49 @@ class PathTests(unittest.TestCase):
         writer.write(root_dir=self.write_path)
         sys.path.append(self.write_path)
 
+        self.check_generated_schemas()
+
+    def check_generated_schemas(self):
+        """ Verify that embedded AVRO schema matches original one """
+        avsc_files = glob.glob(get_joined_path(self.source, "*.avsc"))
+
+        for avsc_file in avsc_files:
+            classname = pathlib.Path(avsc_file).stem
+            module = importlib.import_module("records")
+            clazz = getattr(module, classname)
+            with open(avsc_file, 'r') as f:
+                orig_schema = json.load(f)
+            generated_schema = json.loads(clazz.schema)
+            self.assertEqual(generated_schema, orig_schema, f"Schema mismatch for class: {classname}")
+
     def tearDown(self):
         """ remove all the compiled files"""
         shutil.rmtree(self.write_path + '/records')
+
+    def test_escape_chars(self):
+        """ tests that escape characters are not lost """
+
+        from records import EscapedChars
+        schema = r"""
+    {
+        "type": "record",
+        "name": "EscapedChars",
+        "namespace": "records",
+        "doc": "This is an example \"schema\".",
+        "fields": [
+            {
+                "name": "name",
+                "type": "string"
+            }
+        ]
+    }
+    """
+
+        self.assertEqual(
+            schema,
+            EscapedChars.schema,
+            "Schema should contain escape characters"
+        )
 
     def test_thing(self):
         """ tests that you can serialize and deserialize thing record """
@@ -70,6 +112,36 @@ class PathTests(unittest.TestCase):
             'thing should be able to initialize from setters'
         )
 
+    def test_record_with_numbers(self):
+        """ tests records with numbers work """
+
+        from records import RecordWithNumbers
+
+        data1 = {'booleanValue': True, 'nullableBooleanValue': None, 'intValue': 3, 'longValue': 4000000000,
+                 'floatValue': 1.234E3, 'nullableFloatValue': None, 'doubleValue': 1.234567890123456789E18}
+        data2 = ('{"booleanValue": true, "nullableBooleanValue": null, "intValue": 3, "longValue": 4000000000, '
+                 '"floatValue": 1.234E3, "nullableFloatValue": null, '
+                 '"doubleValue": 1.2345678901234567E18}')
+        data3 = ('{"booleanValue": true, "intValue": 3, "longValue": 4000000000, '
+                 '"floatValue": 1234,'
+                 '"doubleValue": 1234567890123456789}')
+
+        record1 = RecordWithNumbers(data1)
+        record2 = RecordWithNumbers(data2)
+        record3 = RecordWithNumbers(data3)
+
+        self.assertEqual(
+            record1.serialize(),
+            record2.serialize(),
+            'records should be equal'
+        )
+
+        self.assertEqual(
+            record1.serialize(),
+            record3.serialize(),
+            'records should be equal'
+        )
+
     def test_record_with_record(self):
         """ tests nested records work """
 
@@ -79,6 +151,7 @@ class PathTests(unittest.TestCase):
         data1 = {'thing1': {'id': 10}, 'thing2': {'id': 0}}
         data2 = {'thing1': Thing({'id': 10}), 'thing2': Thing({'id': 0})}
         data_with_default = {'thing1': {'id': 10}}
+        invalid_data = {'thing1': None}
 
         record1 = RecordWithRecord(data1)
         record2 = RecordWithRecord(data2)
@@ -94,6 +167,36 @@ class PathTests(unittest.TestCase):
             record1.serialize(),
             record3.serialize(),
             'records should be equal'
+        )
+
+        with self.assertRaises(TypeError):
+            RecordWithRecord(invalid_data)
+
+    def test_record_with_circular_reference(self):
+        """ tests records with circular references """
+
+        from records import RecordWithCircularReference
+        from records import Node
+        import records.nested.Node
+
+        data1 = {'tree': {'id': '0',
+                          'children': [Node({'id': '00', 'parent': Node({'id': '0'})}),
+                                       {'id': '01', 'parent': {'id': '0'},
+                                        'children': [{'id': '010', 'parent': Node({'id': '01'})},
+                                                     {'id': '011',
+                                                      'parent': Node({'id': '01'})}]}]},
+                 'altTree': records.nested.Node({'name': 'x'})}
+        data2 = ('{"tree": {"id": "0", "children": [{"id": "00", "parent": {"id": "0"}}, '
+                 '{"id": "01", "parent": {"id": "0"}, "children": [{"id": "010", "parent": {"id": "01"}}, '
+                 '{"id": "011", "parent": {"id": "01"}}]}]}, "altTree": {"name": "x"}}')
+
+        record1 = RecordWithCircularReference(data1)
+        record2 = RecordWithCircularReference(data2)
+
+        self.assertEqual(
+            record1.serialize(),
+            record2.serialize(),
+            'records with circular references should be equal'
         )
 
     def test_enum(self):
@@ -209,11 +312,19 @@ class PathTests(unittest.TestCase):
         from records import RecordWithArray
         from records import Thing
 
-        data1 = {'things': [{'id': 10}, {'id': 50}], 'numbers': [10, 40], 'things2': [], 'twoDimDoubleArray': [[1.1, 2.2], [3.3, 4.5]], 'threeDimRecordArray': [[[{'id': 10}, {'id': 50}]]]}  # NOQA
-        data2 = {'things': [Thing({'id': 10}), {'id': 50}], 'numbers': [10, 40], 'things2': [], 'twoDimDoubleArray': [[1.1, 2.2], [3.3, 4.5]], 'threeDimRecordArray': [[[Thing({'id': 10}), {'id': 50}]]]}  # NOQA
-        data3 = {'things': [], 'numbers': [], 'things2': [], 'twoDimDoubleArray': [], 'threeDimRecordArray': []}
-        data4 = {'things': [{'id': 10}, {'id': 50}], 'numbers': ['not a long'], 'things2': []}  # NOQA
-        data5 = {'things': [{'id': 'not a long'}, {'id': 50}], 'numbers': [10, 40], 'things2': []}  # NOQA
+        data1 = {'things': [{'id': 10}, {'id': 50}], 'numbers': [10, 40], 'things2': [],
+                 'twoDimDoubleArray': [[1.1, 2.2], [3.3, 4.5]], 'threeDimRecordArray': [[[{'id': 10}, {'id': 50}]]],
+                 'arrayOfUnion': [7, {'id': 8}]}  # NOQA
+        data2 = {'things': [Thing({'id': 10}), {'id': 50}], 'numbers': [10, 40], 'things2': [],
+                 'twoDimDoubleArray': [[1.1, 2.2], [3.3, 4.5]],
+                 'threeDimRecordArray': [[[Thing({'id': 10}), {'id': 50}]]],
+                 'arrayOfUnion': [7, Thing({'id': 8})]}  # NOQA
+        data3 = {'things': [], 'numbers': [], 'things2': [], 'twoDimDoubleArray': [], 'threeDimRecordArray': [],
+                 'arrayOfUnion': []}  # NOQA
+        data4 = {'things': [{'id': 10}, {'id': 50}], 'numbers': ['not a long'], 'things2': [],
+                 'arrayOfUnion': []}  # NOQA
+        data5 = {'things': [{'id': 'not a long'}, {'id': 50}], 'numbers': [10, 40], 'things2': [],
+                 'arrayOfUnion': []}  # NOQA
 
         record1 = RecordWithArray(data1)
         record2 = RecordWithArray(data2)
@@ -227,7 +338,8 @@ class PathTests(unittest.TestCase):
 
         self.assertEqual(
             record3.serialize(),
-            '{"things": [], "numbers": [], "things2": [], "twoDimDoubleArray": [], "threeDimRecordArray": []}'
+            '{"things": [], "numbers": [], "things2": [], "twoDimDoubleArray": [], "threeDimRecordArray": [], '
+            '"arrayOfUnion": []}'
         )
 
         self.assertEqual(
@@ -246,9 +358,11 @@ class PathTests(unittest.TestCase):
 
         from records import RecordWithUnion
         from records import Thing
+        from records.nested import Flavor
 
-        data1 = {'optionalString': 'hello', 'intOrThing': Thing({'id': 2}), 'nullOrThingArray': None}
-        data2 = {'optionalString': 'hello', 'intOrThing': {'id': 2}}
+        data1 = {'optionalString': 'hello', 'intOrThing': Thing({'id': 2}), 'nullOrThingArray': None,
+                 'nullOrEnum': Flavor('CHOCOLATE')}
+        data2 = {'optionalString': 'hello', 'intOrThing': {'id': 2}, 'nullOrEnum': 'CHOCOLATE'}
         data3 = {'optionalString': None, 'intOrThing': 10, 'nullOrThingArray': [{'id': 2}], 'nullOrMap': {'value': 0.1}}
         data4 = {'optionalString': 'hello', 'intOrThing': 'not int or thing'}
 
@@ -257,7 +371,8 @@ class PathTests(unittest.TestCase):
         record3 = RecordWithUnion(data3)
 
         self.assertEqual(
-            '{"optionalString": "hello", "intOrThing": {"id": 2}, "nullOrThingArray": null, "nullOrMap": null}',
+            '{"optionalString": "hello", "intOrThing": {"id": 2}, "nullOrThingArray": null, "nullOrMap": null, '
+            '"nullOrEnum": "CHOCOLATE"}',
             record1.serialize()
         )
 
@@ -283,11 +398,66 @@ class PathTests(unittest.TestCase):
 
         self.assertEqual(
             record3.serialize(),
-            '{"optionalString": null, "intOrThing": 10, "nullOrThingArray": [{"id": 2}], "nullOrMap": {"value": 0.1}}'
+            '{"optionalString": null, "intOrThing": 10, "nullOrThingArray": [{"id": 2}], '
+            '"nullOrMap": {"value": 0.1}, "nullOrEnum": null}'
         )
 
         with self.assertRaises(TypeError):
             RecordWithUnion(data4)
+
+    def test_record_with_geometries(self):
+        """ tests records with union types """
+
+        from records import RecordWithGeometries
+        from records import Point
+        from records import LineString
+        from records import Thing
+
+        data1 = {'geometries': None}
+        data2 = {'geometries': [
+            LineString({"bbox": None, "coordinates": [[-96.41675, 33.968745], [-96.416745, 33.96757]],
+                        "type": "LineString"}),
+            Point({"bbox": None, "coordinates": [-96.41675, 33.968745], "type": "Point"})]}
+        data3 = {'geometries': [
+            {"bbox": None, "coordinates": [[-96.41675, 33.968745], [-96.416745, 33.96757]], "type": "LineString"},
+            {"bbox": None, "coordinates": [-96.41675, 33.968745], "type": "Point"}]}
+        data4 = {'geometries': [Thing({"id": 1})]}
+
+        record1 = RecordWithGeometries(data1)
+        record2 = RecordWithGeometries(data2)
+        record3 = RecordWithGeometries(data3)
+        record4 = RecordWithGeometries(data4)
+
+        self.assertEqual(
+            '{"geometries": null}',
+            record1.serialize()
+        )
+
+        self.assertEqual(
+            record2.serialize(),
+            record3.serialize()
+        )
+
+        self.assertEqual(
+            record2.serialize(),
+            '{"geometries": [{"bbox": null, "coordinates": [[-96.41675, 33.968745], [-96.416745, 33.96757]], '
+            '"type": "LineString"}, {"bbox": null, "coordinates": [-96.41675, 33.968745], "type": "Point"}]}'
+        )
+
+        self.assertIsInstance(
+            record3.geometries[0],
+            LineString
+        )
+
+        self.assertIsInstance(
+            record3.geometries[1],
+            Point
+        )
+
+        self.assertEqual(
+            len(record4.geometries),
+            0
+        )
 
     def test_nested_things(self):
         """ tests that nested things have correct mappings and namespaces """
@@ -326,22 +496,33 @@ class PathTests(unittest.TestCase):
         from records import RecordWithMap
         from records import Thing
 
-        data1 = {'thingMap': [{'string1': {'id': 10}},
+        data1 = {'thingMap': {'string1': {'id': 10},
+                              'string2': Thing({'id': 10})},
+                 'intMap': {'lksdfl': 23},
+                 'thingMap2': {'string1': {'id': 10},
+                               'string2': Thing({'id': 10})},
+                 'thingMap3': {'string1': {'id': 10},
+                               'string2': Thing({'id': 10})},
+                 'arrayMap': {'string1': ['value1', 'value2']},
+                 'arrayOfUnionMap': {
+                     'string1': [Thing({'id': 10}), None]
+                 }
+                 }
+        data2 = {'thingMap': {'string1': {'id': 10},
+                              'string2': Thing({'id': 10})},
+                 'intMap': {'lksdfl': 'NOT A STRING'}}
+        data3 = {'thingMap': [{'string1': {'id': 10}},
                               {'string2': Thing({'id': 10})}],
-                 'intMap': [{'lksdfl': 23}],
-                 'thingMap2': [{'string1': {'id': 10}},
-                               {'string2': Thing({'id': 10})}],
-                 'thingMap3': [{'string1': {'id': 10}},
-                               {'string2': Thing({'id': 10})}]}
-        data2 = {'thingMap': [{'string1': {'id': 10}},
-                              {'string2': Thing({'id': 10})}],
-                 'intMap': [{'lksdfl': 'NOT A STRING'}]}
+                 'intMap': [{'lksdfl': 23}]}
 
         record1 = RecordWithMap(data1)
         record2 = RecordWithMap(record1)
 
         with self.assertRaises(TypeError):
             RecordWithMap(data2)
+
+        with self.assertRaises(TypeError):
+            RecordWithMap(data3)
 
         self.assertEqual(
             record1.serialize(),
@@ -355,77 +536,85 @@ class PathTests(unittest.TestCase):
         from records import Thing
 
         data1 = {
-            'thingMap': [
-                {'string1': {'id': 10}},
-                {'string2': Thing({'id': 10})}
-            ],
-            'nestedThingMap': [
-                {'foo': [
-                    {'string1': {'id': 10}},
-                    {'string2': Thing({'id': 10})}
-                ]},
-                {'bar': [
-                    {'string1': {'id': 10}},
-                    {'string2': Thing({'id': 10})}
-                ]},
-            ],
-            'nestedIntMap': [
-                {'foo': [
-                    {'string1': 10},
-                    {'string2': 11}
-                ]},
-                {'bar': [
-                    {'string1': 10},
-                    {'string2': 11}
-                ]},
-            ],
-            'mappedThingArray': [
-                {'foo': [
+            'thingMap': {
+                'string1': {'id': 10},
+                'string2': Thing({'id': 10})
+            },
+            'nestedThingMap': {
+                'foo': {
+                    'string1': {'id': 10},
+                    'string2': Thing({'id': 10})
+                },
+                'bar': {
+                    'string1': {'id': 10},
+                    'string2': Thing({'id': 10})
+                },
+            },
+            'nestedIntMap': {
+                'foo': {
+                    'string1': 10,
+                    'string2': 11
+                },
+                'bar': {
+                    'string1': 10,
+                    'string2': 11
+                },
+            },
+            'mappedThingArray': {
+                'foo': [
                     {'id': 10},
                     Thing({'id': 20})
-                ]},
-                {'bar': [
+                ],
+                'bar': [
                     {'id': 10},
                     Thing({'id': 20})
-                ]}
-            ]
+                ]
+            }
         }
+
+        serialized_data1 = ('{"thingMap": {"string1": {"id": 10}, "string2": {"id": 10}}, "nestedThingMap": '
+                            '{"foo": {"string1": {"id": 10}, "string2": {"id": 10}}, "bar": {"string1": {"id": 10}, '
+                            '"string2": {"id": 10}}}, "nestedIntMap": {"foo": {"string1": 10, "string2": 11}, "bar": '
+                            '{"string1": 10, "string2": 11}}, "mappedThingArray": {"foo": [{"id": 10}, {"id": 20}], '
+                            '"bar": [{"id": 10}, {"id": 20}]}}')
+
         data2 = {
-            'thingMap': [
-                {'string1': {'id': 10}},
-                {'string2': Thing({'id': 10})}
-            ],
-            'nestedThingMap': [
-                {'foo': [
-                    {'string1': {'id': 10}},
-                    {'string2': Thing({'id': 10})}
-                ]},
-                {'bar': [
-                    {'string1': {'id': 10}},
-                    {'string2': Thing({'id': 10})}
-                ]},
-            ],
-            'nestedIntMap': [
-                {'foo': [
-                    {'string1': 'NOT AN INT'},
-                    {'string2': 11}
-                ]},
-                {'bar': [
-                    {'string1': 10},
-                    {'string2': 11}
-                ]},
-            ],
-            'mappedThingArray': [
-                {'foo': [
+            'thingMap': {
+                'string1': {'id': 10},
+                'string2': Thing({'id': 10})
+            },
+            'nestedThingMap': {
+                'foo': {
+                    'string1': {'id': 10},
+                    'string2': Thing({'id': 10})
+                },
+                'bar': {
+                    'string1': {'id': 10},
+                    'string2': Thing({'id': 10})
+                },
+            },
+            'nestedIntMap': {
+                'foo': {
+                    'string1': 'NOT AN INT',
+                    'string2': 11
+                },
+                'bar': {
+                    'string1': 10,
+                    'string2': 11
+                }
+            },
+            'mappedThingArray': {
+                'foo': [
                     {'id': 10},
                     Thing({'id': 20})
-                ]},
-                {'bar': [
+                ],
+                'bar': [
                     {'id': 10},
                     Thing({'id': 20})
-                ]}
-            ]
+                ]
+            }
         }
+
         record1 = RecordWithNestedMap(data1)
         record2 = RecordWithNestedMap(record1)
 
@@ -435,4 +624,9 @@ class PathTests(unittest.TestCase):
         self.assertEqual(
             record1.serialize(),
             record2.serialize()
+        )
+
+        self.assertEqual(
+            serialized_data1,
+            record1.serialize(),
         )